requirements.txt
flask-marshmallow==0.14.0
pytest==6.2.5
pytest-cov==2.12.1
environs==9.3.3
gunicorn==20.1.0
beautifulsoup4==4.10.0
